% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{computeClassificationMetrics}
\alias{computeClassificationMetrics}
\title{Compute performance metrics for a regression model}
\usage{
computeClassificationMetrics(actual, predicted, metrics = c("Accuracy",
  "Recall", "Precision", "F1_Score"))
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for list
of potential metrics}
}
\description{
Compute performance metrics for a regression model
}
\examples{
data(ptitanic, package = 'rpart.plot')
fit <- rpart(survived ~ ., data = ptitanic)
computeClassificationMetrics(
  actual = ptitanic$survived,
  predicted = predict(fit, type = 'class')
)
}

