% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{computeMultiClassMetrics}
\alias{computeMultiClassMetrics}
\title{Compute performance metrics by class for classification problems}
\usage{
computeMultiClassMetrics(actual, predicted)
}
\arguments{
\item{actual}{vector of actual classes}

\item{predicted}{vector of predicted classes}
}
\description{
Compute performance metrics by class for classification problems
}
\examples{
library(rpart)
fit <- rpart(Species ~ ., data = iris)
pred <- predict(fit, type = 'class')
computeMultiClassMetrics(iris$Species, pred)
}

