% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{computeRegressionMetrics}
\alias{computeRegressionMetrics}
\title{Compute metrics to evaluate performance of a regression model.}
\usage{
computeRegressionMetrics(actual, predicted, metrics = c("MAE", "MAPE",
  "MedianAPE", "RMSE", "RMSLE", "RAE", "R2_Score"))
}
\arguments{
\item{actual}{vector of actual values.}

\item{predicted}{vector of predicted values.}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for
a complete list of metrics that can be used}
}
\description{
Compute metrics to evaluate performance of a regression model.
}
\examples{
mod <- lm(mpg ~ ., data = mtcars)
pred <- predict(mod)
computeRegressionMetrics(mtcars$mpg, pred)
}

