% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{createCoefficientsTable.elnet}
\alias{createCoefficientsTable.elnet}
\title{Create coefficient table from glmnet object}
\usage{
\method{createCoefficientsTable}{elnet}(mod, digits = 3, s = NULL, ...)
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{s}{Value of the penalty parameter lambda at which predictions are
required. Default is the entire sequence used to create the model.}

\item{...}{additional arguments passed to \code{\link{createCoefficientsTable}}}
}
\description{
Create coefficient table from glmnet object
}
