% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{createCoefficientsTable}
\alias{createCoefficientsTable}
\title{Create a table of coefficents}
\usage{
createCoefficientsTable(mod, digits, ...)
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{...}{additional arguments passed to \code{\link{createCoefficientsTable}}}
}
\description{
This is an S3 method that creates a table of coefficients that can be passed
to \code{\link{fdPanelCoefficients}} to display in a dashboard.
}
