% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{fdBoard}
\alias{fdBoard}
\title{Dashboard}
\usage{
fdBoard(header, sidebar, body)
}
\arguments{
\item{header}{A header created by \code{fdHeader}.}

\item{sidebar}{A sidebar created by \code{fdSidebar}.}

\item{body}{A body created by \code{fdBody}.}
}
\description{
This creates a dashboard.
}
\examples{
library(flightdeck)
fdBoard(
  fdHeader(title = 'My Board'),
  fdSidebar(),
  fdBody()
)
}

