% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{fdBox}
\alias{fdBox}
\alias{fdRowBox}
\title{Create a box}
\usage{
fdBox(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 6, height = NULL,
  collapsible = FALSE, collapsed = FALSE, extraBoxClass = NULL,
  extraHeader = NULL)

fdRowBox(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 6, height = NULL,
  collapsible = FALSE, collapsed = FALSE, extraBoxClass = NULL)
}
\arguments{
\item{...}{Contents of the box.}

\item{title}{Optional title.}

\item{footer}{Optional footer text.}

\item{status}{The status of the item This determines the item's background
color.}

\item{solidHeader}{Should the header be shown with a solid color background?}

\item{background}{If NULL (the default), the background of the box will be
white. Otherwise, a color string.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{collapsible}{If TRUE, display a button in the upper right that allows
the user to collapse the box.}

\item{collapsed}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}

\item{extraBoxClass}{extra class to provide to the box.}

\item{extraHeader}{extra header to provide to the box}
}
\description{
Boxes can be used to hold content in the main body of a dashboard.

Syntactic sugar for Box in a Row
}
\seealso{
Other boxes: \code{\link{fdInfoBox}},
  \code{\link{fdValueBox}}

Other boxes: \code{\link{fdInfoBox}},
  \code{\link{fdValueBox}}
}

