% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{fdClassificationMetricsTable}
\alias{fdClassificationMetricsTable}
\title{Display classification metrics in a panel}
\usage{
fdClassificationMetricsTable(actual, predicted, metrics = c("Accuracy",
  "Recall", "Precision", "F1_Score"))
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for list
of potential metrics}
}
\description{
Display classification metrics in a panel
}
\examples{
library(rpart)
data(ptitanic, package = 'rpart.plot')
fit <- rpart(survived ~ ., data = ptitanic)
fdClassificationMetricsTable(
  actual = ptitanic$survived,
  predicted = predict(fit1, type = 'class')
) \%>\%
fdPreview("Classification Metrics")
}

