% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{fdClassificationPerformance}
\alias{fdClassificationPerformance}
\title{Classification Performance By Class}
\usage{
fdClassificationPerformance(actual, predicted, cutoffs = c(0, 0.6, 0.75, 1))
}
\arguments{
\item{actual}{vector of actual classes}

\item{predicted}{vector of predicted classes}

\item{cutoffs}{vector of breaks}
}
\description{
Classification Performance By Class
}
\examples{
data(ptitanic, package = 'rpart.plot')
fit1 = rpart::rpart(survived ~ ., data = ptitanic)
fdClassificationPerformance(
  actual = ptitanic$survived,
  pred = predict(fit1, type = 'class')
) \%>\%
  fdPreview('Classification Performance Metrics')
}

