% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdTidyTable}
\alias{fdTidyTable}
\title{Interactive table of regression coefficients}
\usage{
fdTidyTable(coefTable, digits = 3, barColor = "steelblue")
}
\arguments{
\item{coefTable}{tidy table}

\item{digits}{number of digits to display}

\item{barColor}{bar color}
}
\description{
Interactive table of regression coefficients
}

