% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{fdColumn}
\alias{fdColumn}
\title{Create a column}
\usage{
fdColumn(width, ..., offset = NULL, id = paste0("column-", round(runif(1) *
  10000)))
}
\arguments{
\item{width}{the grid width of the column}

\item{...}{elements to include within the column}

\item{offset}{the number of columns to offset this column}

\item{id}{id for the column}
}
\description{
Create a column
}
