% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{fdHeader}
\alias{fdHeader}
\title{Dashboard header}
\usage{
fdHeader(..., title = NULL, miniTitle = NULL, .list = NULL)
}
\arguments{
\item{...}{Items to put in the header.}

\item{title}{The title to show in the header bar.}

\item{miniTitle}{The title to show in the header bar when sidebar is collapsed.}

\item{.list}{An optional list containing items to put in the header. Same as
the \code{...} arguments, but in list format.}
}
\description{
Dashboard header
}
\examples{
library(htmltools)
myMenu <- fdDropdownMenu(type = "messages",
  fdMessage(
    from = "Sales Dept",
    message = "Sales are steady this month."
  ),
  fdMessage(
    from = "New User",
    message = "How do I register?",
    icon = fdIcon("question"),
    time = "13:45"
  ),
  fdMessage(
    from = "Support",
    message = "The new server is ready.",
    icon = fdIcon("life-ring"),
    time = "2014-12-01"
  )
)


myNotifications <- fdDropdownMenu(type = "notifications",
  fdNotification(
    text = "5 new users today",
    icon = fdIcon("users")
  ),
  fdNotification(
    text = "12 items delivered",
    icon = fdIcon("truck"),
    status = "success"
  ),
  fdNotification(
    text = "Server load at 86\%",
    icon = fdIcon("exclamation-triangle"),
    status = "warning"
  )
)
myHeader <- tagList(myMenu, myNotifications)

if (interactive()){
  fdBoard(
    fdHeader(title = 'Header Menu', myHeader), 
    fdSidebar(), 
    fdBody()
  )
}
}

