% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fdIcon}
\alias{fdIcon}
\title{Create a font icon.}
\usage{
fdIcon(name, class = NULL, lib = "font-awesome", ...)
}
\arguments{
\item{name}{name of the icon}

\item{class}{class to add}

\item{lib}{font library to use}
}
\description{
Create a font icon.
}

