% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fdIconWithPopover}
\alias{fdIconWithPopover}
\title{Icon with tooltip}
\usage{
fdIconWithPopover(name = "info-sign", note = "", class = NULL,
  lib = "glyphicon")
}
\arguments{
\item{name}{name of the icon}

\item{class}{class to add}

\item{lib}{font library to use}
}
\description{
Icon with tooltip
}

