% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{fdMessage}
\alias{fdMessage}
\title{Create a message item to place in a dropdown message menu}
\usage{
fdMessage(from, message, icon = fdIcon("user"), time = NULL, href = NULL)
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{icon}{An icon tag, created by \code{\link{fdIcon}}.}

\item{time}{String representing the time the message was sent. Any string may
be used. For example, it could be a relative date/time like "5 minutes",
"today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45".
If NULL, no time will be displayed.}

\item{href}{An optional URL to link to.}
}
\description{
Create a message item to place in a dropdown message menu
}
\examples{
myMenu <- fdDropdownMenu(type = "messages",
  fdMessage(
    from = "Sales Dept",
    message = "Sales are steady this month."
  ),
  fdMessage(
    from = "New User",
    message = "How do I register?",
    icon = fdIcon("question"),
    time = "13:45"
  ),
  fdMessage(
    from = "Support",
    message = "The new server is ready.",
    icon = fdIcon("life-ring"),
    time = "2014-12-01"
  )
)
fdBoard(fdHeader(myMenu), fdSidebar(), fdBody())
}
\seealso{
\code{\link{fdHeader}} for example usage.

Other menu items: \code{\link{fdNotification}}
}

