% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{fdModal}
\alias{fdModal}
\alias{fdPopover}
\alias{fdTooltip}
\title{Add a modal to an html element}
\usage{
fdModal(x, message, title = NULL, ..., .list = NULL)

fdPopover(x, message, title = NULL, ..., .list = NULL)

fdTooltip(x, message, title = NULL, ..., .list = NULL)
}
\arguments{
\item{x}{html element}

\item{message}{message to dispaly in modal}

\item{title}{title to display in modal}

\item{...}{additional arguments to pass to bootbox/bootstrap. See details.}

\item{.list}{options as a list. makes it easier with programmatic generation.}
}
\description{
Add a modal to an html element
}

