% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{fdNotification}
\alias{fdNotification}
\title{Create a notification}
\usage{
fdNotification(text, icon = fdIcon("warning"), status = "success",
  href = NULL)
}
\arguments{
\item{text}{The notification text.}

\item{icon}{An icon tag, created by \code{\link{fdIcon}}.}

\item{status}{The status of the item This determines the item's background
color.}

\item{href}{An optional URL to link to.}
}
\description{
Create a notification
}
\seealso{
Other menu items: \code{\link{fdMessage}}
}

