% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{fdPage}
\alias{fdPage}
\title{Dashboard Page}
\usage{
fdPage(..., id, display = FALSE)
}
\arguments{
\item{...}{items to put on the page.}

\item{id}{id attribute for the page div.}

\item{display}{boolean indicating if the page is to be displayed by default.}
}
\description{
This function creates a single page.
}

