% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{fdPanelClassificationMetrics}
\alias{computeClassificationMetrics}
\alias{fdPanelClassificationMetrics}
\title{Display performance metrics for a classification model}
\usage{
fdPanelClassificationMetrics(actual, predicted, metrics = c("Accuracy",
  "Recall", "Precision", "F1_Score"))

computeClassificationMetrics(actual, predicted, metrics = c("Precision",
  "Recall", "F1_Score", "Accuracy"))
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for list
of potential metrics}
}
\description{
Display performance metrics for a classification model
}
\examples{
library(flightdeck)
library(rpart)
data(ptitanic, package = 'rpart.plot')
fit <- rpart(survived ~ ., data = ptitanic)
actual <- ptitanic$survived
predicted <- predict(fit, type = 'class')

computeClassificationMetrics(actual = actual, predicted = predicted)

if (interactive()){
  fdPanelClassificationMetrics(actual = actual, predicted = predicted) \%>\%
    fdPreview("Classification Metrics") 
}


}
\seealso{
Other classification: \code{\link{fdPanelClassificationMetricsByClass}}
}

