% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{fdPanelClassificationMetricsByClass}
\alias{computeClassificationMetricsByClass}
\alias{fdPanelClassificationMetricsByClass}
\title{Display performance metrics by class for a classification model}
\usage{
fdPanelClassificationMetricsByClass(actual, predicted, vertical = TRUE,
  cutoffs = c(0, 0.6, 0.75, 1))

computeClassificationMetricsByClass(actual, predicted, vertical = TRUE)
}
\arguments{
\item{actual}{vector of actual classes.}

\item{predicted}{vector of predicted classes.}

\item{vertical}{boolean indicating if panel should display measures in columns.}

\item{cutoffs}{vector of cutoffs to use for deciding bar colors.}
}
\description{
Display performance metrics by class for a classification model
}
\examples{
library(rpart)
data(ptitanic, package = 'rpart.plot')
fit <- rpart(survived ~ ., data = ptitanic)
actual <- ptitanic$survived
predicted <- predict(fit, type = 'class')

computeClassificationMetrics(actual = actual, predicted = predicted)

if (interactive()){
  fdPanelClassificationMetrics(actual = actual, predicted = predicted) \%>\%
    fdPreview("Classification Metrics") 
}


}
\seealso{
Other classification: \code{\link{fdPanelClassificationMetrics}}
}

