% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdPanelCoefficients}
\alias{fdPanelCoefficients}
\title{Displaying regression coefficients}
\usage{
fdPanelCoefficients(mod, digits = 3, barColor = "steelblue")
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{barColor}{bar color}
}
\description{
Displaying regression coefficients
}
\examples{
library(flightdeck)
data(diamonds, package = 'ggplot2')
if (interactive()){
  lm(price ~ ., data = diamonds) \%>\%
    fdPanelCoefficients \%>\%
    fdPreview  
}

}

