% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdPanelImportance}
\alias{fdPanelImportance}
\title{Interactive variable importance table}
\usage{
fdPanelImportance(mod, digits = 2, barColor = "steelblue")
}
\description{
Interactive variable importance table
}
\examples{
library(rpart)
data(ptitanic, package = 'rpart.plot')
if (interactive()){
  rpart(survived ~ ., data = ptitanic) \%>\%
    fdPanelImportance \%>\%
    fdPreview
}
}

