% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-diagnostics.R
\name{fdPanelRegressionDiagnostics}
\alias{fdPanelRegressionDiagnostics}
\alias{fdPlotsRegressionDiagnostics}
\title{Display a panel of diagnostic plots for regression models}
\usage{
fdPanelRegressionDiagnostics(mod)

fdPlotsRegressionDiagnostics(mod)
}
\arguments{
\item{mod}{a regression model object}
}
\description{
Display a panel of diagnostic plots for regression models
}
\examples{
mod <- lm(mpg ~ ., data = mtcars)
fdPanelRegressionDiagnostics(mod) \%>\% fdPreview(wrap = 'none')
}
