% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelRegressionMetrics}
\alias{computeRegressionMetrics}
\alias{fdPanelRegressionMetrics}
\title{Display performance metrics for a regression model}
\usage{
fdPanelRegressionMetrics(actual, predicted, metrics = c("MAE", "MAPE",
  "MedianAPE", "RMSE", "RMSLE", "RAE", "R2_Score"))

computeRegressionMetrics(actual, predicted, metrics = c("MAE", "MAPE",
  "MedianAPE", "RMSE", "RMSLE", "RAE", "R2_Score"))
}
\arguments{
\item{actual}{vector of actual values.}

\item{predicted}{vector of predicted values.}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for
a complete list of metrics that can be used.}
}
\description{
Display performance metrics for a regression model

Compute metrics to evaluate performance of a regression model
}
\examples{
library(flightdeck)
data(cars)
reg <- lm(log(dist) ~ log(speed), data = cars)
actual <- cars$dist
predicted <- exp(reg$fitted.values)
fdPanelRegressionMetrics(actual, predicted) \%>\%
  fdPreview("Regression Metrics")
}

