% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPanels.R
\name{fdPanelRegressionResiduals}
\alias{fdPanelRegressionResiduals}
\title{Panel displaying model residuals}
\usage{
fdPanelRegressionResiduals(mod, digits = 4)
}
\arguments{
\item{mod}{model object}
}
\description{
Panel displaying model residuals
}
\examples{
library(flightdeck)
mod <- lm(mpg ~ ., data = mtcars)
if (interactive()){
  mod \%>\% fdPanelRegressionResiduals \%>\% fdPreview
}
}

