% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelRegressionScatterplot}
\alias{fdPanelRegressionScatterplot}
\title{Dispaly scatterplot of actual vs. predicted values for a regression model.}
\usage{
fdPanelRegressionScatterplot(actual, predicted)
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}
}
\description{
Dispaly scatterplot of actual vs. predicted values for a regression model.
}
\examples{
library(flightdeck)
fit <- lm(mpg ~ ., data = mtcars)
actual <- mtcars$mpg
predicted <- predict(fit)
if (interactive()){
  fdPanelRegressionScatterplot(actual, predicted) \%>\% 
    fdPreview
}
}

