% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fdPlotClassificationPerformance}
\alias{fdPlotClassificationPerformance}
\title{Interactive performance plots for classification models.}
\usage{
fdPlotClassificationPerformance(perf, digits = 3, title = NULL)
}
\arguments{
\item{perf}{an object of class \code{\link[ROCR]{performance}}}

\item{digits}{number of digits to display}

\item{title}{optional title}
}
\description{
Interactive performance plots for classification models.
}
\examples{
library(flightdeck)
library(ROCR)
data(ROCR.simple)
pred <- prediction(ROCR.simple$predictions, ROCR.simple$labels)
perf <- performance(pred, 'tpr', 'fpr')


rocChart <- fdPlotClassificationPerformance(
  performance(pred, "tpr", "fpr"), digits = 2
)
precisionRecallChart <- fdPlotClassificationPerformance(
  performance(pred, "prec", "rec"), digits = 2
)

# Example 1
fdTabsetPanel(selected = makeHtmlId('ROC Chart'),
  fdTabPanel('ROC Chart', rocChart),
  fdTabPanel('Precision vs. Recall', precisionRecallChart)
) \%>\%
  fdColumn(width = 12) \%>\%
  fdPreview(wrap = 'row')
}

