% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fdPlotConditionalDensity}
\alias{fdPlotConditionalDensity}
\title{Interactive Conditional Density Plot}
\usage{
fdPlotConditionalDensity(x, y, title = NULL, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a "factor" interpreted to be the dependenty variable}

\item{title}{optional title for the plot}

\item{...}{additional arguments to pass to \code{\link[graphics]{cdplot}}}
}
\description{
Interactive Conditional Density Plot
}
\examples{
library(flightdeck)
# Example 1: NASA space shuttle o-ring failures
fail <- factor(c(2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 2, 1, 2, 1, 1, 1,
  1, 2, 1, 1, 1, 1, 1),
  levels = 1:2, labels = c("no", "yes")
)
temperature <- c(
  53, 57, 58, 63, 66, 67, 67, 67, 68, 69, 70, 70,
  70, 70, 72, 73, 75, 75, 76, 76, 78, 79, 81
)
# Conditional Density Plot
# cdplot(fail ~ temperature)
if (interactive()){
  fdPlotConditionalDensity(temperature, fail)
}

# Example 2: Iris
if (interactive()){
  fdPlotConditionalDensity(iris$Petal.Length, iris$Species, xlab = 'Petal.Length')
}
}

