% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-glmnet.R
\name{fdPlotCvGlmnet}
\alias{fdPlotCvGlmnet}
\title{Interactive plot for cv.glmnet object.}
\usage{
fdPlotCvGlmnet(x, sign.lambda = NULL, title = NULL, ...)
}
\arguments{
\item{x}{fitted "cv.glmnet" object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative
if sign.lambda=-1}

\item{title}{plot title.}

\item{...}{additional arguments. not currently used}
}
\description{
This plots MSE vs Log Lambda
}
\examples{
library(glmnet)
library(flightdeck)
data(QuickStartExample)
cvfit <- cv.glmnet(x, y)

if (interactive()){
  #' Interctive profiles plot
  fdPlotCvGlmnet(cvfit)
}
}
