% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-glmnet.R
\name{fdPlotGlmnet}
\alias{fdPlotGlmnet}
\title{Interactive coefficient profile plot for glmnet objects.}
\usage{
fdPlotGlmnet(x, xvar = c("norm", "lambda", "dev"), title = NULL, ...)
}
\arguments{
\item{x}{a glmnet model object}

\item{xvar}{What is on the X-axis. "norm" plots against the L1-norm of the
coefficients, "lambda" against the log-lambda sequence, and "dev" against
the percent deviance explained.}

\item{title}{plot title.}

\item{...}{extra arguments passed to every series being plotted}
}
\description{
Interactive coefficient profile plot for glmnet objects.
}
\examples{
library(glmnet)
library(flightdeck)
data(QuickStartExample)
fit <- glmnet(x, y)

if (interactive()){
  #' Interctive profiles plot
  fdPlotGlmnet(fit)
  
  # Use lambda as xvar
  fdPlotGlmnet(fit, xvar = 'lambda')
  
  # Don't show legend
  fdPlotGlmnet(fit, showlegend = FALSE)
}
}

