% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{fdPlotly}
\alias{fdPlotly}
\title{Create an interactive plot using plotly.}
\usage{
fdPlotly(data, layout = NULL, config = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{list of traces to plot.}

\item{layout}{list consisting of the layout to use for the plot.}

\item{config}{list consisting of configuration to use for the plot.}

\item{width}{width of the plot.}

\item{height}{height of the plot.}
}
\description{
This function is a lightweight wrapper around plotly.js. For complete
documentation, please refer to \url{https://plot.ly/javascript/reference/}.
}
\examples{
# Example 1: Scatterplot
dat <- list(
  x = runif(100),
  y = runif(100),
  mode = 'markers',
  type = 'scatter'
)
layout <- list(margin = list(t = 40))
config <- list(displaylogo = FALSE, displayModeBar = FALSE)

if (interactive()){
  fdPlotly(
    data = list(dat),
    layout = layout,
    config = config,
    height = 325
  )  
}


# Example 2: Line and Scatter Plot
# https://plot.ly/javascript/line-and-scatter/
trace1 = list(
  x =  c(1, 2, 3, 4),
  y =  c(10, 15, 13, 17),
  mode =  'markers',
  type =  'scatter'
)

trace2 = list(
  x =  c(2, 3, 4, 5),
  y =  c(16, 5, 11, 9),
  mode =  'lines',
  type =  'scatter'
)

trace3 = list(
  x =  c(1, 2, 3, 4),
  y =  c(12, 9, 15, 12),
  mode =  'lines+markers',
  type =  'scatter'
)

if (interactive()){
  fdPlotly(
    data = list(trace1, trace2, trace3)
  )
}
}
