% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{fdPreview}
\alias{fdPreview}
\title{Preview a widget in the dashboard.}
\usage{
fdPreview(widget, title = deparse(substitute(widget)), wrapBox = TRUE, ...)
}
\arguments{
\item{widget}{widget to preview}

\item{title}{title to display}
}
\description{
Preview a widget in the dashboard.
}
\examples{
library(flightdeck)
fdSimpleTable(mtcars) \%>\% fdPreview(title = 'Simple Table')
}

