% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdRegressionMetricsTable}
\alias{fdRegressionMetricsTable}
\title{Display regression metrics in a panel}
\usage{
fdRegressionMetricsTable(actual, predicted, metrics = c("MAE", "MAPE",
  "MedianAPE", "RMSE", "RMSLE", "RAE", "R2_Score"))
}
\arguments{
\item{actual}{vector of actual values.}

\item{predicted}{vector of predicted values.}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for
a complete list of metrics that can be used}
}
\description{
Display regression metrics in a panel
}
\examples{
data(cars)
reg <- lm(log(dist) ~ log(speed), data = cars)
fdRegressionMetricsTable(
  actual = cars$dist,
  predicted = exp(reg$fitted.values)
) \%>\%
fdPreview("Regression Metrics")
}

