% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdRegressionScatterplot}
\alias{fdRegressionScatterplot}
\title{Scatterplot of actual vs. predicted values for a regression model}
\usage{
fdRegressionScatterplot(actual, predicted)
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}
}
\description{
Scatterplot of actual vs. predicted values for a regression model
}
\examples{
fit <- lm(mpg ~ ., data = mtcars)
actual <- mtcars$mpg
predicted <- predict(fit)
fdRegressionScatterplot(actual, predicted) \%>\% fdPreview
}

