% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{fdRender}
\alias{fdRender}
\title{Render dashboard in Alteryx Composer}
\usage{
fdRender(x, libdir = NULL, nOutput = 3, debug = FALSE,
  selfcontained = TRUE)
}
\arguments{
\item{x}{an object of class htmlwidget or shiny.tag.}

\item{libdir}{directory to copy js/css dependencies to.}

\item{nOutput}{connection to pipe output back into Alteryx.}

\item{debug}{only for use in development}

\item{selfcontained}{boolean indicating if html generated should be selfcontained.}
}
\description{
Render dashboard in Alteryx Composer
}

