% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPanels.R
\name{fdResidualsPanel}
\alias{fdResidualsPanel}
\title{Panel displaying model residuals}
\usage{
fdResidualsPanel(mod, digits = 4)
}
\arguments{
\item{mod}{model object}
}
\description{
Panel displaying model residuals
}
\examples{
mod <- lm(mpg ~ ., data = mtcars)
mod \%>\% fdResidualsPanel \%>\% fdPreview
}

