% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{fdSidebarMenu}
\alias{fdMenuItem}
\alias{fdMenuSubItem}
\alias{fdSidebarMenu}
\title{Sidebar menu}
\usage{
fdSidebarMenu(...)

fdMenuItem(text, ..., icon = NULL, pageName = NULL, href = "#")

fdMenuSubItem(text, href = NULL, pageName = NULL,
  icon = fdIcon("angle-double-right"))
}
\arguments{
\item{...}{Items to put in the sidebar. It can include \code{\link{fdMenuSubItem}}}

\item{text}{Text to show for the menu item.}

\item{icon}{Icon to display against the text.}

\item{pageName}{Id of the page that this menu item will activate}

\item{href}{Link address.}
}
\description{
This creates a sidebar menu which mainly consists of
\code{\link{fdMenuItem}} and  \code{\link{fdMenuSubItem}}
}

