% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdSimpleTable}
\alias{fdSimpleTable}
\title{Create a simple html table}
\usage{
fdSimpleTable(x, class = "table table-striped")
}
\arguments{
\item{x}{data frame to be displayed as a html table}
}
\description{
Create a simple html table
}
\examples{
fdSimpleTable(mtcars) \%>\% fdPreview
}

