% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdTable}
\alias{fdTable}
\title{Syntactic sugar to create a html table from a list of columns}
\usage{
fdTable(..., class = "table table-striped")
}
\arguments{
\item{...}{columns for the table.}

\item{class}{css calss for table.}
}
\description{
Syntactic sugar to create a html table from a list of columns
}

