% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{fdTabsetPanel}
\alias{fdTabContent}
\alias{fdTabContentItem}
\alias{fdTabNav}
\alias{fdTabNavItem}
\alias{fdTabPanel}
\alias{fdTabsetPanel}
\title{Make a tabset panel}
\usage{
fdTabsetPanel(..., selected = NULL)

fdTabPanel(title, ..., id = title)

fdTabNav(...)

fdTabContent(...)

fdTabNavItem(..., id = NULL, active = FALSE)

fdTabContentItem(..., id = NULL, active = FALSE)
}
\arguments{
\item{...}{elements to put in. see details.}

\item{selected}{id of the active tab}

\item{active}{boolean indicating if tab is active}

\item{...}{elements to put in the panel}

\item{name}{unique name to identify tab}
}
\description{
Make a tabset panel

Make a tab panel
}

