% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fdTitleWithPopover}
\alias{fdTitleWithPopover}
\title{Text with tooltip that has a popover}
\usage{
fdTitleWithPopover(text, note = "", name = "info-sign", class = NULL,
  lib = "glyphicon", ...)
}
\arguments{
\item{note}{note to add to the popover.}

\item{name}{name of the icon}

\item{class}{class to add}

\item{lib}{font library to use}

\item{title}{text of the title.}
}
\description{
Text with tooltip that has a popover
}

