% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{fdTour}
\alias{fdTour}
\title{Add a tour}
\usage{
fdTour(steps, options, button = NULL, width = 30, height = 10)
}
\arguments{
\item{steps}{list of steps for the tour}

\item{options}{options for the tour}

\item{button}{boolean indicating if a tour button should be displayed}

\item{width}{width of the tour button}

\item{height}{height of the tour button}
}
\description{
This widget creates a tour on a web page.
}

