% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdVariableImportanceTable}
\alias{fdVariableImportanceTable}
\title{Interactive variable importance table}
\usage{
fdVariableImportanceTable(mod, digits = 2, barColor = "steelblue")
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{barColor}{bar color}
}
\description{
Interactive variable importance table
}
\examples{
library(rpart)
data(ptitanic, package = 'rpart.plot')
m1 <- rpart(survived ~ ., data = ptitanic)
fdVariableImportanceTable(m1) \%>\% fdPreview
}

