% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeHtmlId}
\alias{makeHtmlId}
\title{Make a valid html id from a string}
\usage{
makeHtmlId(x)
}
\arguments{
\item{x}{string to convert into a valid html id.}
}
\description{
Make a valid html id from a string
}

