% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe functions}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Use the pipe function, \code{\%>\%} to turn function composition
into a series of imperative statements.
}

