% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyLite.R
\name{plotlyLite}
\alias{plotlyLite}
\title{A light wrapper around plotly.js}
\usage{
plotlyLite(data, layout = NULL, config = NULL, width = NULL,
  height = NULL)
}
\description{
This function is a lightweight wrapper around plotly.js. For complete
documentation, please refer to \url{https://plot.ly/javascript/reference/}
}
\examples{
 d = data.frame(
   x = 1:5,
   y = (1:5)*2
 )
 plotlyLite(data = list(d))
}

