% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{fdClassificationPerformance}
\alias{computeClassificationMetrics}
\alias{fdClassificationPerformance}
\title{Compute metrics to evaluate performance of a classification model.}
\usage{
fdClassificationPerformance(actual, predicted, cutoffs = c(0, 0.6, 0.75, 1))

computeClassificationMetrics(actual, predicted, metrics = c("Accuracy",
  "Recall", "Precision", "F1_Score"))
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}

\item{cutoffs}{vector of breaks}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for list
of potential metrics}
}
\description{
Compute metrics to evaluate performance of a classification model.
}
\examples{
data(ptitanic, package = 'rpart.plot')
fit <- rpart(survived ~ ., data = ptitanic)
computeClassificationMetrics(
  actual = ptitanic$survived,
  predicted = predict(fit, type = 'class')
)
}

