% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{computeRegressionMetrics}
\alias{computeRegressionMetrics}
\title{Compute performance metrics for a regression model}
\usage{
computeRegressionMetrics(actual, predicted, metrics = c("MAE", "MAPE",
  "MedianAPE", "RMSE", "RMSLE", "RAE", "R2_Score"))
}
\arguments{
\item{actual}{vector of actual values}

\item{predicted}{vector of predicted values}

\item{metrics}{vector of metrics to compute. See \code{\link{MLmetrics}} for list
of potential metrics}
}
\description{
Compute performance metrics for a regression model
}
\examples{
mod <- lm(mpg ~ ., data = mtcars)
pred <- predict(mod)
computeRegressionMetrics(mtcars$mpg, pred)
}

