% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{fdBoard}
\alias{fdBoard}
\title{Create a dashboard}
\usage{
fdBoard(header, sidebar, body, theme = "skin-blue-light")
}
\arguments{
\item{header}{A header created by \code{fdHeader}.}

\item{sidebar}{A sidebar created by \code{fdSidebar}.}

\item{body}{A body created by \code{fdBody}.}

\item{theme}{A string indicating theme to use for the dashboard.}
}
\description{
This function creates a dashboard.
}
\examples{
#' Example 1
if (interactive()){
  fdBoard(
    fdHeader(),
    fdSidebar(),
    fdBody()
  ) 
}

#' Example 2
if (interactive()){
  fdBoard(
    fdHeader(title = 'My Board'),
    fdSidebar(),
    fdBody(
      h3("Hello World!")
    )
  ) 
}
}

