% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdCoefTable}
\alias{fdCoefTable}
\title{Interactive table of regression coefficients}
\usage{
fdCoefTable(mod, digits = 3, barColor = "steelblue")
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{barColor}{bar color}
}
\description{
Interactive table of regression coefficients
}
\examples{
data(diamonds, package = 'ggplot2')
lm(price ~ ., data = diamonds) \%>\%
  fdCoefTable \%>\%
  fdPreview
}

