% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{fdHeader}
\alias{fdHeader}
\title{Dashboard header}
\usage{
fdHeader(..., title = NULL, miniTitle = NULL, .list = NULL)
}
\arguments{
\item{...}{Items to put in the header.}

\item{title}{The title to show in the header bar.}

\item{miniTitle}{The title to show in the header bar when sidebar is collapsed.}

\item{.list}{An optional list containing items to put in the header. Same as
the \code{...} arguments, but in list format.}
}
\description{
Dashboard header
}

