% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{fdHeader}
\alias{fdHeader}
\title{Create the header of the dashboard}
\usage{
fdHeader(..., title = NULL, miniTitle = NULL, titleWidth = NULL,
  .list = NULL)
}
\arguments{
\item{...}{Elements to in the dashboard header.}

\item{title}{The title to display in the header bar.}

\item{miniTitle}{The title to display in the header bar when the sidebar is
collapsed.}

\item{titleWidth}{The width of the title in pixels.}

\item{.list}{An optional list containing elements to put in the header. These
are the identical to the \code{...} arguments, but provided in a list
format. This is useful when the elements going into the header are
programatically generated using \code{lapply} calls}
}
\description{
Create the header of the dashboard
}
\examples{
library(htmltools)
myMenu <- fdDropdownMenu(type = "messages",
  fdMessage(
    from = "Sales Dept",
    message = "Sales are steady this month."
  ),
  fdMessage(
    from = "New User",
    message = "How do I register?",
    icon = fdIcon("question"),
    time = "13:45"
  ),
  fdMessage(
    from = "Support",
    message = "The new server is ready.",
    icon = fdIcon("life-ring"),
    time = "2014-12-01"
  )
)


myNotifications <- fdDropdownMenu(type = "notifications",
  fdNotification(
    text = "5 new users today",
    icon = fdIcon("users")
  ),
  fdNotification(
    text = "12 items delivered",
    icon = fdIcon("truck"),
    status = "success"
  ),
  fdNotification(
    text = "Server load at 86\%",
    icon = fdIcon("exclamation-triangle"),
    status = "warning"
  )
)
myHeader <- tagList(myMenu, myNotifications)

if (interactive()){
  fdBoard(
    fdHeader(title = 'Header Menu', myHeader), 
    fdSidebar(), 
    fdBody()
  )
}
}

