% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdPanelCoefficients}
\alias{fdPanelCoefficients}
\title{Display regression coefficients}
\usage{
fdPanelCoefficients(mod, digits = 3, barColor = "steelblue", ...)
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{barColor}{bar color}

\item{...}{additional arguments passed to \code{\link{createCoefficientsTable}}}
}
\description{
Display regression coefficients
}
\examples{
library(flightdeck)

#' Example 1: Linear Regression Model
data(diamonds, package = 'ggplot2')
panel1 <- lm(price ~ ., data = diamonds) \%>\%
  fdPanelCoefficients \%>\%
  fdPreview  

if (interactive()) {
  panel1
}

#' Example 2: GlmNet Model
library(glmnet)
data(QuickStartExample)
panel2 <- glmnet(x, y) \%>\%
  fdPanelCoefficients(s = 0.3) \%>\%
  fdPreview

if (interactive()){
  panel2
}
}
