% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{fdPanelHistogram}
\alias{fdPanelHistogram}
\title{Display a histogram and table of summary statistics}
\usage{
fdPanelHistogram(x, digits = 4, plotTitle = "Histogram", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{number of digits to display in the summary statistics table.}

\item{plotTitle}{title of the histogram.}

\item{...}{extra arguments (not used currently).}
}
\description{
Display a histogram and table of summary statistics
}
\examples{
library(flightdeck)
if (interactive()){
  runif(1000) \%>\% 
    fdPanelHistogram(plotTitle = 'Histogram of Residuals') \%>\%
    fdPreview
}
}

