% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdPanelImportance}
\alias{fdPanelImportance}
\title{Panel displaying variable importance.}
\usage{
fdPanelImportance(mod, digits = 2, barColor = "steelblue")
}
\arguments{
\item{mod}{model object}

\item{digits}{number of digits to display}

\item{barColor}{bar color}
}
\description{
Panel displaying variable importance.
}
\examples{
library(rpart)
data(ptitanic, package = 'rpart.plot')
if (interactive()){
  rpart(survived ~ ., data = ptitanic) \%>\%
    fdPanelImportance \%>\%
    fdPreview
}
}

