% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelMetrics}
\alias{fdPanelMetrics}
\title{Display a panel of metrics}
\usage{
fdPanelMetrics(x)
}
\arguments{
\item{x}{a data frame containing metrics to display.}
}
\description{
Display a panel of metrics
}

