% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelRegressionResiduals}
\alias{fdPanelRegressionResiduals}
\title{Display model residuals for a linear regression model}
\usage{
fdPanelRegressionResiduals(mod, digits = 4, plotTitle = "Histogram", ...)
}
\arguments{
\item{mod}{lm model object with a residuals method}

\item{digits}{number of digits to display in the summary statistics table}

\item{plotTitle}{title of the histogram.}

\item{...}{additional arguments. not used currently.}
}
\description{
Display model residuals for a linear regression model
}
\examples{
library(flightdeck)
mod <- lm(mpg ~ ., data = mtcars)
if (interactive()){
  mod \%>\% 
    fdPanelRegressionResiduals(plotTitle = 'Histogram of Residuals') \%>\%
    fdPreview
}
}
