% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelRegressionResiduals}
\alias{fdPanelRegressionResiduals}
\title{Displaying model residuals for a regression model}
\usage{
fdPanelRegressionResiduals(mod, digits = 4, plotTitle = "Histogram", ...)
}
\arguments{
\item{mod}{model object with a residuals method}

\item{digits}{number of digits to display in the summary statistics table}

\item{plotTitle}{title of the histogram.}
}
\description{
Displaying model residuals for a regression model
}
\examples{
library(flightdeck)
mod <- lm(mpg ~ ., data = mtcars)
if (interactive()){
  mod \%>\% 
    fdPanelRegressionResiduals(plotTitle = 'Histogram of Residuals') \%>\%
    fdPreview
}
}

