% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fdPanelRegressionResiduals}
\alias{fdPanelRegressionResiduals}
\title{Displaying model residuals for a regression model}
\usage{
fdPanelRegressionResiduals(mod, digits = 4)
}
\arguments{
\item{mod}{model object}
}
\description{
Displaying model residuals for a regression model
}
\examples{
library(flightdeck)
mod <- lm(mpg ~ ., data = mtcars)
if (interactive()){
  mod \%>\% fdPanelRegressionResiduals \%>\% fdPreview
}
}

