% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fdPlotConfusionMatrix}
\alias{fdPlotConfusionMatrix}
\title{Interactive Confusion Matrix}
\usage{
fdPlotConfusionMatrix(x, class = "table table-bordered")
}
\arguments{
\item{x}{confusion matrix with row and column names representing classes}

\item{class}{html class to style the confusion matrix table}
}
\description{
Interactive Confusion Matrix
}
\examples{
library(flightdeck)
library(rpart)
mod <- rpart(Species ~ ., data = iris)
outcome <- as.character(attr(mod, 'ylevels')[mod$y])
pred <- predict(mod, type = "class")
cmat <- as.data.frame.matrix(table(outcome, pred))

if (interactive()){
  fdRowBox(width = 6,
    fdPlotConfusionMatrix(cmat),
    title = 'Confusion Matrix',
    footer = paste(
      'Numbers in parantheses indicate \% of observations', 
      'correctly classified. Hover for more details'
    ) 
  ) \%>\% 
  fdPreview(wrap = 'none')
}


}

