% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{fdPreview}
\alias{fdPreview}
\alias{fdPreviewBoard}
\title{Preview a widget in the dashboard.}
\usage{
fdPreview(widget, title = deparse(substitute(widget)), wrap = "rowbox", ...,
  wrapBox = NULL)

fdPreviewBoard(widget, title = deparse(substitute(widget)), wrap = "rowbox",
  ..., wrapBox = NULL)
}
\arguments{
\item{widget}{widget to preview}

\item{title}{title to display}

\item{wrap}{string indicating if the widget is to be wrapped in a 'row',
'box', 'rowbox' or 'none'
\code{\link{fdRowBox}}}

\item{...}{additional elements to pass to \code{\link{fdBody}}.}

\item{wrapBox}{boolean indicating if we should set wrap = 'rowbox'}
}
\description{
This function helps preview a widget in the dashboard with or without the
header and sidebar.
}
\examples{
library(flightdeck)
fdSimpleTable(mtcars) \%>\% fdPreview(title = 'Simple Table')
fdSimpleTable(mtcars) \%>\% fdPreviewBoard(title = 'Simple Table')
}
