% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{fdPreview}
\alias{fdPreview}
\alias{fdPreviewBoard}
\title{Preview a widget in the dashboard.}
\usage{
fdPreview(widget, title = deparse(substitute(widget)), wrapBox = TRUE, ...)

fdPreviewBoard(widget, title = deparse(substitute(widget)), wrapBox = TRUE,
  ...)
}
\arguments{
\item{widget}{widget to preview}

\item{title}{title to display}

\item{wrapBox}{boolean indicating if the widget is to be wrapped in an
\code{\link{fdRowBox}}}

\item{...}{additional elements to pass to \code{\link{fdBody}}.}
}
\description{
This function helps preview a widget in the dashboard with or without the
header and sidebar.
}
\examples{
library(flightdeck)
fdSimpleTable(mtcars) \%>\% fdPreview(title = 'Simple Table')
fdSimpleTable(mtcars) \%>\% fdPreviewBoard(title = 'Simple Table')
}

