% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{fdSidebar}
\alias{fdSidebar}
\title{Dashboard sidebar}
\usage{
fdSidebar(...)
}
\arguments{
\item{...}{Items to put in the sidebar.}
}
\description{
This creates a sidebar for the dashboard
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  fdBoard(
    fdHeader(title='Sidebar'),
    fdSidebar(
      fdSidebarMenu(
        fdMenuItem("Page1",icon = fdIcon("table"),
          fdMenuSubItem('Page1A', pageName = 'page1')
        ),
        fdMenuItem("Page2",icon=fdIcon("bar-chart"), pageName = 'page2')
      ) 
    ),
    fdBody(
      fdPage(p('Page 1'), id = 'page1', display = TRUE),
      fdPage(p('Page 2'),id = 'page2')
    )
  )
}
}
\seealso{
\code{\link{fdSidebarMenu}}
}

