% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fdSimpleTable}
\alias{fdSimpleTable}
\alias{fdTable}
\title{Create a table}
\usage{
fdSimpleTable(x, class = "table table-striped")

fdTable(..., class = "table table-striped")
}
\arguments{
\item{x}{data frame to be displayed as a html table.}

\item{class}{html class to add to the table.}

\item{...}{columns for the table.}
}
\description{
Create a table
}
\details{
The \code{\link{fdTable}} function is just syntactic sugar for
\code{\link{fdSimpleTable}}
}
\examples{
fdSimpleTable(mtcars) \%>\% fdPreview
}

