% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{fdTabsetPanel}
\alias{fdTabsetPanel}
\alias{fdTabPanel}
\alias{fdTabNav}
\alias{fdTabContent}
\alias{fdTabNavItem}
\alias{fdTabContentItem}
\title{Make a tabset panel}
\usage{
fdTabsetPanel(..., selected = NULL, .list = NULL)

fdTabPanel(title, ..., id = makeHtmlId(title))

fdTabNav(...)

fdTabContent(...)

fdTabNavItem(..., id = NULL, active = FALSE)

fdTabContentItem(..., id = NULL, active = FALSE)
}
\arguments{
\item{...}{elements to put in. see details.}

\item{selected}{id of the active tab}

\item{.list}{elements to put in as a list.}

\item{title}{title to use.}

\item{id}{html id to add.}

\item{active}{boolean indicating if tab is active.}
}
\description{
Make a tabset panel

Make a tab panel
}
